refactor(#3037): extract WS broadcast bus to crate::eventbus (kill service backflow)
dispatched_sessions (a service) reached into crate::server::ws::emit_event /
emit_batched_event — 5 of the remaining service→server backflow references.

Move the broadcast-bus primitives (BroadcastBus/BroadcastEvent/PendingEvent/
BroadcastTx/BatchBuffer/new_broadcast/emit_event/emit_batched_event/
spawn_batch_flusher) into a new crate-root module `crate::eventbus`, which sits
below both `server` and `services` in the dependency graph. The axum
ws_handler/handle_socket stay in `crate::server::ws`, which re-exports the
primitives it still uses so all existing `crate::server::ws::*` call sites
resolve unchanged. dispatched_sessions now emits via `crate::eventbus::*`.

Pure relocation: logic/orderings/visibility-of-behavior unchanged
(as_ws_message lifted to pub(crate) for the cross-module handler). Backflow -5.
